Fix overflow handling in ft_atoi

+++ ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: stakada <stakada@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/17 04:20:37 by stakada           #+#    #+#             */
-/*   Updated: 2024/05/02 06:20:09 by stakada          ###   ########.fr       */
+/*   Updated: 2024/11/12 01:28:06 by stakada          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,9 +29,9 @@ int	ft_atoi(const char *str)
 	}
 	while (*str >= '0' && *str <= '9')
 	{
-		if (sign == 1 && result > (LONG_MAX - (*str - '0')) / 10)
+		if (sign == 1 && result > (INT_MAX - (*str - '0')) / 10)
 			return ((int)LONG_MAX);
-		else if (sign == -1 && -(result) < (LONG_MIN + (*str - '0')) / 10)
+		else if (sign == -1 && -(result) < (INT_MIN + (*str - '0')) / 10)
 			return ((int)LONG_MIN);
 		result = result * 10 + (*str - '0');
 		str++;
